refactor(client): drop cross-workspace session relation UI
Session trees are worktree-scoped now, so the browser no longer resolves or
displays relations that cross a workspace: the parent-location resolver, the
"Go to parent session" action, the children-elsewhere meta line, and the
optimistic child count for sessions created elsewhere are gone, together with
the SessionInfo fields the server stopped producing.

Orphan rows (a recorded parent that is not in this listing) keep the dimmed
child glyph and now state only that the parent is unavailable here.
