+++ src/client/src/controllers/sessionController.ts
@@ -14,7 +14,6 @@ import { isSessionActive } from "../../../shared/activity";
 import type { PromptAttachmentDelivery, SessionNotificationInboxEvent, SessionStartupProgressEvent } from "../../../shared/apiTypes";
 import { InMemorySessionSelectionMemory, markSessionArchived, markSessionsArchived, selectPreferredSession, selectionAfterArchivingSession, selectionAfterArchivingSessions, shouldDeselectAfterArchivedCollapse, type SessionSelectionMemory } from "./sessionSelection";
 import { selectedMachineId, type GetState, type SetState, type UpdateUrl } from "./types";
-import { sessionPathsEqual } from "../sessionPaths";
 import { TrailingRefreshCoordinator } from "./trailingRefreshCoordinator";
 
 const MESSAGE_PAGE_SIZE = 100;
@@ -1283,14 +1282,9 @@ export class SessionController {
 
   private applyCreatedSession(session: SessionInfo) {
     const state = this.getState();
-    // A session created in another workspace is not listed here, but it may be a
-    // child of one that is: keep that parent's cross-workspace child count live
-    // instead of leaving it stale until the next listing.
-    if (state.selectedWorkspace?.path !== session.cwd) {
-      const patch = childElsewhereCountPatch(state, session);
-      if (patch !== undefined) this.setState(patch);
-      return;
-    }
+    // Session trees are worktree-scoped, so a session created elsewhere has no
+    // bearing on this listing.
+    if (state.selectedWorkspace?.path !== session.cwd) return;
     if (state.sessions.some((candidate) => candidate.id === session.id)) return;
     const machineId = selectedMachineId(state);
     if (this.hasPendingStartFor(session.cwd, machineId)) {
@@ -1761,27 +1755,6 @@ function bulkFailureMessages(failures: readonly SessionBulkFailure[]): string[] 
 
 function errorMessage(error: unknown): string {
   return error instanceof Error ? error.message : String(error);
-}
-
-/**
- * State patch incrementing the cross-workspace child count of the listed parent
- * of `createdSession`, or undefined when that parent is not in view.
- *
- * Exported for testing: it encodes when a session created in another workspace
- * is relevant to the current listing at all.
- */
-export function childElsewhereCountPatch(state: AppState, createdSession: SessionInfo): Pick<Partial<AppState>, "sessions"> | undefined {
-  const parentPath = createdSession.parentSessionPath;
-  if (parentPath === undefined || parentPath === "") return undefined;
-  // The created session's parent path and the listed session's own path reach the
-  // browser from different server producers, so they are compared tolerantly.
-  const isParent = (candidate: SessionInfo) => sessionPathsEqual(candidate.path, parentPath);
-  if (!state.sessions.some(isParent)) return undefined;
-  return {
-    sessions: state.sessions.map((candidate) => isParent(candidate)
-      ? { ...candidate, childSessionsElsewhere: (candidate.childSessionsElsewhere ?? 0) + 1 }
-      : candidate),
-  };
 }
 
 function sessionMessageCountPatch(state: AppState, sessionId: string, messageCount: number | undefined): Pick<Partial<AppState>, "sessions" | "selectedSession"> {
